SENSEI/S8N/SUPERSEIZ - verified contract 
There was quite a lot of issue with verifying the contracts. 
We used default Remix to deploy and for some reason that contract was not possible to verify later, so we ended up redeploying the tokens to a new address. 
This is something we recommend the core team to look at as many people use default remix settings to deploy. 

This PR is updating the contract addresses (to the now verified addresses) and moving the 3 assets to pacific-1-evm.